huaweicloud-sdk-core/huaweicloudsdkcore/http/future_session.py
# coding: utf-8
"""
 Licensed to the Apache Software Foundation (ASF) under one
 or more contributor license agreements.  See the NOTICE file
 distributed with this work for additional information
 regarding copyright ownership.  The ASF licenses this file
 to you under the Apache LICENSE, Version 2.0 (the
 "LICENSE"); you may not use this file except in compliance
 with the LICENSE.  You may obtain a copy of the LICENSE at

     http://www.apache.org/licenses/LICENSE-2.0

 Unless required by applicable law or agreed to in writing,
 software distributed under the LICENSE is distributed on an
 "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 KIND, either express or implied.  See the LICENSE for the
 specific language governing permissions and limitations
 under the LICENSE.
"""

from requests import Session


class FutureSession(Session):
    def __init__(self, session, executor):
        super(FutureSession, self).__init__()
        self._session = session
        self._executor = executor

    def request(self, *args, **kwargs):
        return self._executor.submit(self._session.request, *args, **kwargs)
