huaweicloud-sdk-core/requirements.txt
requests-toolbelt >= 0.10.1
simplejson >= 3.18.0
six >= 1.16.0
PyYAML >= 5.4.1
certifi >= 2022.12.7; python_version > '2.7'
futures >= 3.3.0; python_version == '2.7'
configparser >= 4.0.2; python_version == '2.7'
